SW-40 - Validate: Documentation and script updates (#73)
* SW-40 - Validate: Documentation and script updates

Shell script has been updated to try and resolve the case where
JAVA_HOME is not set. It utilizes the `which java` command to try and
find a java installation on the target platform. If it is still cannot
be found at this point, then an error message is issued.