Issue #55: Update POM to use latest PDS4-JParser snapshot
Bug is fixed in PDS4-JParser and the way it handles the VICAR dependency. No idea why the software wasn't catching the error, but I couldn't seem to track down the reason why it was failing validation but no errors were thrown. Punting that down the road in the event another issue arises.