Update field validator to allow empty fields for delimited tables
Updates made per #9 sufficed for character tables, but caused an issue for delimited tables.

This update allows for blank fields, but still catches issues when a space character is used instead.

According the PDS DSV Standard in the SR (SR 4C.1) "Leading and trailing spaces in a field are considered part of the field" so that is not an issue.

Resolves #137