Hacky temp workaround for schematron bug
Per #175 and #183, we have hit an snag where we have to upgrade Saxon
in order to prevent memory leak, but we still need to workaround this Saxon
bug where it is unable to handle `parent::node()` functionality.

This update hacks the SchemaTransformer to replace any `parent::node()` with a different
method to get the same result, prior to creating the tranformer object.

Refs #182