Update compare method to use BigDecimal.scale (#577)
Fix 1
-----
Original compare method had some confusions between precision and scale. We should be using scale for handling
doubles with many significant digits.

> The precision is the number of digits in the unscaled value. For instance, for the number 123.45, the precision returned is 5.

> If zero or positive, the scale is the number of digits to the right of the decimal point. If negative, the unscaled value of the number is multiplied by ten to the power of the negation of the scale. For example, a scale of -3 means the unscaled value is multiplied by 1000.

Additionally, we need to be able to handle scale appropriately. Previous functionality would decrease the scale to the least between the min/max value and the actual value. This obviously doesn't make sense because when we do this, comparisons like this will be invalid:

actual value: 0.0014194006
scaled actual value: 0.0014194
max value from label: 0.0014194

as you can see, if we compare the max value from the label to the scaled actual value, they are equal. but in reality they are not. this should raise an error.

Fix 2
-----
that being said, we can't always trust the full scaled values that are calculated because doubles can introduce false precision. if we just change the algorithm to use whichever is the largest scale (aka number of decimal places) we were taking values that sometimes introduced false precision. thanks for our regression test in #435, we were able to see that the solution is not as simple as increasing the scale. so after some reading, it looks like using the BigDecimal(String) initializer as much as possible, and avoiding calculations/conversions to double in memory is the best way to avoid false precision. there are still some possible issues with false precision in the code, but this should resolve most of them.

resolves #529

References:
* https://blogs.oracle.com/javamagazine/post/four-common-pitfalls-of-the-bigdecimal-class-and-how-to-avoid-them
* https://stackoverflow.com/questions/35435691/bigdecimal-precision-and-scale