added new error
When the length of the record read from the length given in the XML does not match that reading by delimiter, toss an error. Used RECORD_LENGTH_MISMATCH as that seemed appropriate rather than creaating a new one. Cleaned up the end of line check too. Not unit test pushed with this one as the results from the given data are mirky at best.

Lastly, requires changes to pds4-jparser as well.