add the actual check
Verify the file extension or suffix matches those expected for the given encoding type. An enumeration (table) is used to make the XML encoding type to a list of allowed extensions for the encoding type. The enumeration is the only thing that needs changed when new encoding types are added or list of the extensions is expanded.