Throw WARNING when encountering non-standard file extensions for common document encoding types (#635)
* add test first

* partial fix to see if it breaks anything

* clean up

* fix Document_File

* fix Document_File

* add the actual check

Verify the file extension or suffix matches those expected for the given encoding type. An enumeration (table) is used to make the XML encoding type to a list of allowed extensions for the encoding type. The enumeration is the only thing that needs changed when new encoding types are added or list of the extensions is expanded.

---------

Co-authored-by: Al Niessner <Al.Niessner@xxx.xxx>
Co-authored-by: Jordan Padams <33492486+jordanpadams@users.noreply.github.com>