change index to ID
In both Array and Tables if a name is given then that will be used as well as the index to identify it. Not sure that matches what is really desired, but it seems more meaningful than just the name or just the index value. I had seen them many of times and had no idea that it was the index in the XML file.