Handle bundle/collection files as input with bundle/collection rule (#656)
* add unit test first - reuse part of 597

* touch up feature line

* get rid of $ in feature file

* change to TargetExaminer

Found the existing code used TargetExaminer to determine if a target is a bundle or collection so got rid of the same checks in AggregateManager and moved the code to use the check for better uniform behavior.

* change to TargetExaminer

Found the existing code used TargetExaminer to determine if a target is a bundle or collection so got rid of the same checks in AggregateManager and moved the code to use the check for better uniform behavior.

* change to TargetExaminer

Found the existing code used TargetExaminer to determine if a target is a bundle or collection so got rid of the same checks in AggregateManager and moved the code to use the check for better uniform behavior.

* change to TargetExaminer

Found the existing code used TargetExaminer to determine if a target is a bundle or collection so got rid of the same checks in AggregateManager and moved the code to use the check for better uniform behavior.

* handle bundle/collection files

Upgraded AggregationManager to handle bundle and collection XML files instead of just directories. Still end up with lots of skipped files if they exist in the directories because that is hard to fix. Would have to process the older versions like when a directory is specified to build up the list of older versions or do some other strange processing to know they are old versions at the place where the skipped files pop up. Seems like the skipped messages should remain because they tell the user that validate knows they are there but not part of the collection/bundle that they were specifically told to process.

* fix some weird merge bug

---------

Co-authored-by: Al Niessner <Al.Niessner@xxx.xxx>