fix broken names
The PDF error messages where using the wrong names. Corrected the error messages to use the correct names to reduce confusion. This ticket and subsequent investigation were because the names being reported were the first or parent item in the Document_Edition and not the current name. Now it clearly shows the errors are related to the correct files. Once the reporter saw the correct files with erros they fixed the PDFs.