fix the bug
Through many of the tickets things had purposely been turned off due to a large amount of noise #368. When the reporting was turned on for this ticket got 8 warnings when the expectation was one or two. Found that the reference for some were not for this bundle. Added further filtering that the reference had to be contained in this bundle as well.

Found many of the lid/lidvid checks where contains() instead of startsWith(). Change to startsWith() to prevent some messed up thing where there is a match in the middle of the string when it clearly must startWith() for correctness. This too may have some cleanup to the output.