finished the duplication checks
Parse the aggregations to query back for the lids that reference it. Then compress those lids in a set then report duplicates if the set is bigger than 1. This allows many lidvids to point at the same thing if they have the same lid.

Tested using the m2020 pixl data from harvest#143 after duplicating one element for the check to find. Runs successfully.