rewrite the report generation
Moved to a vistor pattern more or less. Used the existing structures so the visiting is done via abstract calls from the parent. All of the structure and keywords are now provided through the Report class and the implementers of it just serialize what they are given into their given formats. They all now have the same content.