fixed memory leak(s)
There were many. First, it turns out that hashing on a URL or URI is a bad idea. When doing profiling, the tables that hashed on these objects were tremendously big (thousands). Changeing them hash (key) on .toString() reduced the tables to 12 items. Seemed more reasonable.

Second problem was the schematron were holding their past information in memory and then recycling them. resulted in 10s of MB per label processed being stored in each of the schematron validators. Doing a reset() on the validator when doen, cleared that problem.

Third, the schematron validators were holding onto document frangments that translated into 1 MB per label (give or take as it dependend on the label size). To get rid of these, need to clean up (garbage collect) the schematron validators. This is going to cost in CPU cycles per label because the validator has to be created from scratch each time. However, long jobs will not fail because of out of memory.

As part of (3), changed from caching the transform to caching the actual document. Hence it is loaded once from file or network and then used repeatedly to create new schematron validators. In this way, bet benefit of not reloadeing a million times, but pay to create validators. Not too much slower. Doing about 300 labels every 15 minutes currently. Think it was faster yesterday at maybe 400 every 15 minutes but not positive. Never wrote it down.