dup check taking foreever
In this case the inventory file had million plus rows. The way duplicates were being computed was to iterate over all entries then iterate over all entries. Hence the loop was a million million. So, changed the code so that it iterates once. It completes the million in many milliseconds now.