update to SDK v2
Moved to using registry-common and harvest for the tools they provide. harvest provides the configuration reader while the registry-common provides the interface to any opensearch.

Updated the logic to use the new tools which made the unit tests invalid. Removed all of the unit test code for RI since it is no longer applicable.

The logic of RI has remained. What has changed is how the interface to opensearch is used.