rewind jaxb update
The changes from there to bigger than 2x is significant. The interesting part is how sporadic the errors are. The bottom line is this:

Caused by: java.lang.ClassNotFoundException: com.sun.xml.internal.bind.v2.ContextFactory
	at java.base/java.lang.ClassLoader.findClass(ClassLoader.java:723)
	at java.base/java.lang.ClassLoader.loadClass(ClassLoader.java:592)
	at java.base/java.lang.ClassLoader.loadClass(ClassLoader.java:525)
	at javax.xml.bind.ServiceLoaderUtil.nullSafeLoadClass(ServiceLoaderUtil.java:122)
	at javax.xml.bind.ServiceLoaderUtil.safeLoadClass(ServiceLoaderUtil.java:155)
	at javax.xml.bind.ContextFinder.newInstance(ContextFinder.java:276)

It seems that some configuration or help needs to be given to fill in the now missing class. Pulling back this change because its use is limited and thus its real security risk is minimal. The files it reads are provided by the user and thus trusted.