fixes on last bugfix
Updated the pom to use a patched version of xml-resolver. Not sure why it is not published by the original authors, but it did fix the test killing eclipse.

Moved the last set of problems to bottom of the test stack so that {reportDir}/catalog.xml will be for those three last troublesome tests.