cleaned up expectatiosn
Three of the tests were trying to download or access network resources that were not available during the recreation of the feature files. These tests had some erroneous expectations because of the external lack of resources. While the fault is clear it does demonstrate a weakness of the tests that they can fail for reasons beyond the control of the validate repo.