fix up datetime processing
Needed an new problem type for this since they are not numbers as it is currently meant.

Used the DateTimeValidator to do the conversions from string to Instant. Turns out that the string needs some normalization and the pattern being used to convert also needs love depending on the value. Little more complicated than I would have liked.

Removed some dead code from FieldValueValidator because it would just make the error messages confusing and wrong.

Bulk of the update is in SpecialConstantChecker but is straight forward. The why is to support the ticket in general.