extend using the radix
The table field processing made a big assumption assuming that all cells would be base 10. It is true except for when they are declared otherwise with ASCII_NUMERIC_BASE*. Now do checks for the base and declare a radix which defaults to 10. Now all processing uses the given radix unless specifically overriden in the special constants with 0x, 0o, 0b or the other declarations.