1090: count products as skipped when --skip-product-validation (#1200)
* gov.nasa.pds.validate.report.Report

It is the sole place where failed, ignored, passed, and skipped are accumulated.  Since files skipped from --skip-product-validation no longer count as passed, had to change how the total items are computed as well as if a task goes in the passed or skipped bucket.

* test

Since how passed and skipped are counted, had to update some of the test expectations.